news_research_project/requirements.txt
streamlit
python-dotenv
langchain-groq
langchain-huggingface
langchain-community
langchain-core
langchain-text-splitters
faiss-cpu
unstructured
sentence-transformers